requirements.txt
djangorestframework~=3.14.0
Django~=4.2.10
django-cors-headers~=4.3.1
psycopg2-binary==2.9.9
weasyprint==61.2
PyJWT==2.8.0